Add option to force kill debugpy adapter for Zed
This option is required as a workaround for the Zed editor that
(incorrectly?) does not terminate the debugpy-adapter when finishing the
debug session, as per issue
https://github.com/zed-industries/zed/issues/34266.

If you use this option with Zed then note that when debugging you can't
immediately press the "rerun session" button. You must press "detach" or
"terminate all threads" then wait a second or so before pressing "rerun
session" to give time for the adapter to be killed and the port to be
re-available.

Use the following command line to use `debugpy-run` with Zed:

    $ debugpy-run -EK myprog.py -- [--myprog-args ..]

Of course you must also configure your Zed `debug.json` appropriately
for remote attach.