Kill adapter on first run if using that option
If you forget to specify the -K option when using zed then a stray
adapter will be left running so you can just restart debugpy-run with the
-K option and it will kill that process off before starting. Previously
debugpy-run was only killing the adapter between runs.

Also, if you need the -K option then you must be using zed so now we
also force the -E option to not use the VS Code extension.